Refactored project layout for Visual Studio, removed unnecessary headers, and removed ignored files.
