--- README.md
@@ -1,2 +1,0 @@
-# c-clickerrpg
-A CMD-based clicker game I'm working on in C++
